Store the cons pairs in an array
Representing a (x . y) pair as two consecutive elements in an array is
much more efficient than allocating a struct with references to the
elements. The new representation only (re)allocates memory when the
array has to be resized, compared to the older one allocating on every
cons operation.

In addition, the new implementation allows a natural representation of
a nil value terminating a list. Previously, we would have to store the
nil value as a singleton in the state to allow referencing it. Now, we
can simply copy it whenever needed.